+++ tests/integration/optlyplugins/userprofileservice/no_op_ups.go
@@ -14,16 +14,20 @@
  * limitations under the License.                                           *
  ***************************************************************************/
 
-package optlyplugins
+package userprofileservice
 
 import "github.com/optimizely/go-sdk/pkg/decision"
 
-// LookupErrorUserProfileService represents a user profile service with lookup error
-type LookupErrorUserProfileService struct {
+// NoOpUserProfileService represents a user profile service with save and lookup error
+type NoOpUserProfileService struct {
 	NormalUserProfileService
 }
 
 // Lookup is used to retrieve past bucketing decisions for users
-func (s *LookupErrorUserProfileService) Lookup(userID string) decision.UserProfile {
+func (s *NoOpUserProfileService) Lookup(userID string) decision.UserProfile {
 	return decision.UserProfile{}
+}
+
+// Save is used to save bucketing decisions for users
+func (s *NoOpUserProfileService) Save(userProfile decision.UserProfile) {
 }
