+++ tests/integration/optlyplugins/userprofileservice/save_error_ups.go
@@ -14,20 +14,17 @@
  * limitations under the License.                                           *
  ***************************************************************************/
 
-package optlyplugins
+package userprofileservice
 
-import "github.com/optimizely/go-sdk/pkg/decision"
+import (
+	"github.com/optimizely/go-sdk/pkg/decision"
+)
 
-// NoOpUserProfileService represents a user profile service with save and lookup error
-type NoOpUserProfileService struct {
+// SaveErrorUserProfileService represents a user profile service with save error
+type SaveErrorUserProfileService struct {
 	NormalUserProfileService
 }
 
-// Lookup is used to retrieve past bucketing decisions for users
-func (s *NoOpUserProfileService) Lookup(userID string) decision.UserProfile {
-	return decision.UserProfile{}
-}
-
 // Save is used to save bucketing decisions for users
-func (s *NoOpUserProfileService) Save(userProfile decision.UserProfile) {
+func (s *SaveErrorUserProfileService) Save(userProfile decision.UserProfile) {
 }
